pyproject.toml
[tool.poetry]
name = "make.env"
version = "0.0.5"
description = "Infuses GNU make with the ability to read .env files"
authors = ["smac89 <nobleman.code@gmx.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/smac89/make.env"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Topic :: Software Development :: Build Tools",
]
keywords = ["make", ".env", "Makefile", "automation", "buildtool"]

[tool.poetry.dependencies]
python = "^3.6"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
mypy = "^0.931"

[tool.poetry.scripts]
"make.env" = "make_env:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
